backend/requirements.txt
Flask
openai
python-dotenv
requests
flask_cors